requirements.txt
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
fonttools==4.53.0
idna==3.7
ipython==8.25.0
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.4
pillow==10.3.0
prompt_toolkit==3.0.46
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.1
wcwidth==0.2.13

scipy~=1.14.0